refactor(common): split utilities into focused modules
